Rename bundle processing modules and split out pipeline validation
Rename the package's bare modules to the <subject>_<kind> convention used
across afft.deployment: types.py, builders.py, registry.py and runners.py
become pipeline_types.py, pipeline_builders.py, processor_registry.py and
pipeline_runners.py.

Move validate_pipeline out of the runner module into pipeline_validators.py.
It is a precondition callers invoke, not something run_pipeline performs for
them, and sharing a module with run_pipeline obscured that. The package now
holds one module per phase of a pipeline's life: types, builders, validators,
runners.

PipelineStepError stays with run_pipeline; the validators raise plain
ValueError. No signature, exception type or exported name changes, so the
package's public API is unchanged.
