+++ src/afft/bundle_processing/__init__.py
@@ -1,23 +1,23 @@
 """Package for bundle-in/bundle-out processing pipelines."""
 
-from .types import (
+from .pipeline_types import (
     Pipeline as Pipeline,
     PipelineConfig as PipelineConfig,
     PipelineProcessor as PipelineProcessor,
     PipelineStep as PipelineStep,
     PipelineStepConfig as PipelineStepConfig,
 )
-from .registry import (
+from .processor_registry import (
     PipelineProcessorRegistry as PipelineProcessorRegistry,
     RegisteredProcessor as RegisteredProcessor,
     default_registry as default_registry,
     register_processor as register_processor,
 )
-from .builders import build_pipeline as build_pipeline
-from .runners import (
+from .pipeline_builders import build_pipeline as build_pipeline
+from .pipeline_validators import validate_pipeline as validate_pipeline
+from .pipeline_runners import (
     PipelineStepError as PipelineStepError,
     run_pipeline as run_pipeline,
-    validate_pipeline as validate_pipeline,
 )
 
 # Imported for their registration side effects: a processor in an unimported
