+++ src/afft/bundle_processing/pipeline_validators.py
@@ -1,0 +1,32 @@
+"""Checks a resolved pipeline against the bundle it is to run on. Callers
+run these before `run_pipeline`, which does not validate for them."""
+
+from collections.abc import Set
+
+from .pipeline_types import Pipeline
+
+
+def validate_pipeline(pipeline: Pipeline, available: Set[str]) -> None:
+    """
+    Check that every step's inputs are available when that step runs.
+
+    Arguments
+    ---------
+    pipeline: The resolved steps, in run order.
+    available: Keys present before the first step -- the input bundle's.
+
+    Raises
+    ------
+    ValueError: If a step names an input key that no earlier step produces
+        and the input bundle does not hold.
+    """
+    produced: set[str] = set(available)
+    for index, step in enumerate(pipeline):
+        for name, key in step.inputs.items():
+            if key not in produced:
+                raise ValueError(
+                    f"step {index} ({step.processor_key}) reads {name}="
+                    f"{key!r}, which the input bundle does not hold and no "
+                    f"earlier step produces"
+                )
+        produced.add(step.output)
