+++ tests/test_bundle_processing_processor_registry.py
